Don't repeat download if you can and cleanup devmux

+++ pkg/download-wallpapers.sh
@@ -11,6 +11,8 @@ laskyline="https://nextcloud.cmrn.xyz/s/rSMFMJxcoJTJyGY/download"
 declare -A backgrounds=( [losangelesnight.jpg]="$losangelesnight"  [hongkongnight.jpg]="$hongkongnight" [laskyline.jpg]="$laskyline")
 
 for key in "${!backgrounds[@]}"; do
-    echo "downloading $key"
-    wget -O "$key" "${backgrounds[$key]}"
+    if [ ! -e "$DEST/$key" ]; then
+        echo "downloading $key"
+        wget -O "$key" "${backgrounds[$key]}"
+    fi
 done
